v1.6.0 for package rename, update to 6.0
